refactor(Gallery/amp): mac didn't pickup the case sensitive renaming :(
