+++ package.json
@@ -38,7 +38,7 @@
     "downshift": "1.30.0",
     "expo": "23.0.4",
     "formik": "0.11.4",
-    "react-helmet": "^5.2.0",
+    "react-helmet": "5.2.0",
     "react-overlays": "0.8.3",
     "yup": "0.24.1"
   },
