+++ src/components/Gallery/amp/Gallery.js
