+++ src/components/Gallery/amp/GalleryContainer.js
