+++ src/components/Gallery/amp/index.js
