+++ src/components/Image/amp/Image.amp.story.js
