+++ src/components/Image/amp/Image.js
