+++ src/components/Image/amp/Img.js
