+++ src/components/Image/amp/index.js
